unified the locks, and more between kai and judit
