+++ neuromapp/keyvalue/benchmarks/task.h
@@ -1,5 +1,5 @@
 /*
- * Neuromapp - cstep.cpp, Copyright (c), 2015,
+ * Neuromapp - task.h, Copyright (c), 2015,
  * Timothee Ewart - Swiss Federal Institute of technology in Lausanne,
  * timothee.ewart@epfl.ch,
  * All rights reserved.
@@ -19,8 +19,8 @@
  * License along with this library.
  */
 
-#ifndef MAPP_BENCHMARK_H
-#define MAPP_BENCHMARK_H
+#ifndef MAPP_TASK_H
+#define MAPP_TASK_H
 
 #include <vector>
 #include <numeric>
@@ -37,93 +37,12 @@
 
 #include "utils/mpi/timer.h"
 
+//forward declaration
 template<keyvalue::selector S>
-class benchmark{
-public:
-    typedef typename keyvalue::trait_meta<S>::meta_type meta_type; // extract the good meta
-    /** \fun benchmark(std::size_t choice, std::size_t qmb = 4096)
-        \brief compute the total number of compartment (2.5 = 2.5 MB per neuron, 350 compartiment per neuron)
-        4096 MB, 25% of the memory of a compute node of the BG/Q
-     */
-    benchmark(keyvalue::argument const& args = keyvalue::argument()):a(args){
-        s = args.voltages_size();
-        int cg_size = s / args.cg();
-        int first_size = cg_size + (s % args.cg());
-
-        g = keyvalue::group<meta_type>(cg_size);
-        g.push_back(keyvalue::nrnthread(first_size));
-
-        for (int i = 1; i < args.cg(); i++)
-            g.push_back(keyvalue::nrnthread(cg_size));
-    }
-
-    /** \fun get_group() const
-        \brief get the group i.e. the memory */
-    keyvalue::group<meta_type> const& get_group() const{
-        return g;
-    }
-
-    /** \fun get_args() const
-        \brief return the argument */
-    keyvalue::argument const & get_args() const {
-        return a;
-    }
-
-private:
-    /** memory for the bench */
-    keyvalue::group<meta_type> g;
-    /** reference on the arguments structure */
-    keyvalue::argument const & a;
-    /** correspond to the total number of compartement */
-    std::size_t s;
-};
+class benchmark;
 
 template<keyvalue::selector S>
-keyvalue::statistic run_loop(benchmark<S> const& b){
-    typedef typename keyvalue::trait_meta<S>::meta_type meta_type; // get the good meta type
-    // extract the group of memory
-    keyvalue::group<meta_type> const& g = b.get_group(); //git it to the group
-    keyvalue::argument const& a = b.get_args();
-
-    // build the needed function in function of the backend
-    typename keyvalue::trait_meta<S>::keyvalue_type kv;
-    kv.async() = a.async();
-
-    // the timer
-    mapp::timer t;
-    // go to dodo
-    int comp_time_us = 100 * a.usecase() * 1000;
-
-    // keep time trace
-    std::vector<double> vtime;
-    vtime.reserve(1024);
-
-    // these two loops should be merge
-    for (float st = 0; st < a.st(); st += a.md()) {
-        for (float md = 0; md < a.md(); md += a.dt()) {
-            usleep(comp_time_us);
-
-            t.tic();
-
-            #pragma omp parallel for
-            for (int cg = 0; cg < a.cg(); cg++)
-                kv.insert(g.meta_at(cg));
-
-
-            #pragma omp parallel for
-            for (int cg = 0; cg < a.cg(); cg++)
-                kv.wait(g.meta_at(cg));
-
-            t.toc();
-            vtime.push_back(t.time());
-        }
-    }
-
-    return keyvalue::statistic(a,vtime);
-}
-
-template<keyvalue::selector S>
-std::pair<keyvalue::statistic,keyvalue::statistic> run_task(benchmark<S> const& b){
+std::pair<keyvalue::statistic,keyvalue::statistic> task(benchmark<S> const& b){
     typedef typename keyvalue::trait_meta<S>::meta_type meta_type; // get the good meta type
     // extract the group of memory and the arguement
     keyvalue::group<meta_type> const& g = b.get_group(); //get the group
