+++ neuromapp/utils/omp/lock.h
@@ -29,7 +29,7 @@
 #ifdef _OPENMP
 #include <omp.h>
 
-namespace queueing {
+namespace mapp {
 
 class omp_lock{
 private:
@@ -60,7 +60,7 @@ public:
 }
 #endif
 
-namespace queueing {
+namespace mapp {
 
 class dummy_lock{
 public:
